move telemetry code into its own module.
