+++ packages/teleport/src/boot.tsx
@@ -17,30 +17,11 @@ limitations under the License.
 import ReactDOM from 'react-dom';
 import React from 'react';
 
-import { Resource } from '@opentelemetry/resources';
-import { SemanticResourceAttributes } from '@opentelemetry/semantic-conventions';
-import { WebTracerProvider } from '@opentelemetry/sdk-trace-web';
-import { registerInstrumentations } from '@opentelemetry/instrumentation';
-import {
-  ConsoleSpanExporter,
-  BatchSpanProcessor,
-} from '@opentelemetry/sdk-trace-base';
-import { OTLPTraceExporter } from '@opentelemetry/exporter-trace-otlp-http';
-import { DocumentLoadInstrumentation } from '@opentelemetry/instrumentation-document-load';
-import { XMLHttpRequestInstrumentation } from '@opentelemetry/instrumentation-xml-http-request';
-import { FetchInstrumentation } from '@opentelemetry/instrumentation-fetch';
-import { B3Propagator } from '@opentelemetry/propagator-b3';
-import { UserInteractionInstrumentation } from '@opentelemetry/instrumentation-user-interaction';
-import {
-  CompositePropagator,
-  W3CTraceContextPropagator,
-} from '@opentelemetry/core';
-import { ZoneContextManager } from '@opentelemetry/context-zone';
-
 import history from 'teleport/services/history';
 
 import Teleport from './Teleport';
 import TeleportContext from './teleportContext';
+import { instantiateTelemetry } from './telemetry-boot';
 import cfg from './config';
 
 // apply configuration received from the server
@@ -49,44 +30,12 @@ cfg.init(window['GRV_CONFIG']);
 // use browser history
 history.init();
 
-// Optionally register automatic instrumentation libraries
-registerInstrumentations({
-  instrumentations: [
-    new DocumentLoadInstrumentation(),
-    new UserInteractionInstrumentation(),
-    new XMLHttpRequestInstrumentation(),
-    new FetchInstrumentation(),
-  ],
-});
+const enableTelemetry =
+  localStorage.getItem('enable-telemetry') === 'true' ? true : false;
 
-const resource = Resource.default().merge(
-  new Resource({
-    [SemanticResourceAttributes.SERVICE_NAME]: 'web-ui',
-    [SemanticResourceAttributes.SERVICE_VERSION]: '0.1.0',
-  })
-);
-
-const provider = new WebTracerProvider({
-  resource: resource,
-});
-
-provider.addSpanProcessor(new BatchSpanProcessor(new ConsoleSpanExporter()));
-provider.addSpanProcessor(
-  new BatchSpanProcessor(
-    new OTLPTraceExporter({
-      timeoutMillis: 15000,
-      url: 'https://localhost:8080/v1/webapi/traces', // url is optional and can be omitted - default is http://localhost:4318/v1/traces
-      concurrencyLimit: 10, // an optional limit on pending requests
-    })
-  )
-);
-
-provider.register({
-  contextManager: new ZoneContextManager(),
-  propagator: new CompositePropagator({
-    propagators: [new B3Propagator(), new W3CTraceContextPropagator()],
-  }),
-});
+if (enableTelemetry) {
+  instantiateTelemetry();
+}
 
 const teleportContext = new TeleportContext();
